+++ tests/test_panels_twitter.py
@@ -128,3 +128,16 @@ class TestMetaPreviewTwitterAdminView(TestCase, WagtailTestUtils):
                 "default_image": "",
             },
         )
+
+    def test_title_and_description_trunctation(self):
+        twitter_settings = TwitterSettings(self.twitter_page)
+
+        meta_settings.META_PREVIEW_TWITTER_TITLE_FIELDS = "title"
+        meta_settings.META_PREVIEW_TWITTER_DESCRIPTION_FIELDS = "search_description"
+
+        self.twitter_page.title = "This title exeeds the threshold recommended for twitter titles and should be capped at max length"
+        self.twitter_page.search_description = "This text is exactly 200 characters long, designed to convey the idea of counting characters. The goal is to show how we can be precise with word choices, making sure to reach the exact character count! Everything outside this is capped."
+        self.twitter_page.save()
+
+        self.assertEqual(len(twitter_settings.get_title()), 70)
+        self.assertEqual(len(twitter_settings.get_description()), 200)