+++ tests/app/settings.py
@@ -68,7 +68,7 @@ TEMPLATES = [
     {
         "BACKEND": "django.template.backends.django.DjangoTemplates",
         "APP_DIRS": True,
-        "DIRS": "templates",
+        "DIRS": ["templates"],
         "OPTIONS": {
             "context_processors": [
                 "django.template.context_processors.debug",