+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Changed
 - Improve developement instructions (@marteinn)
+- Make sure title and descriptions are truncated if too long (@marteinn)
 
 ### Fixed
 - Fix broken image preview #12 (@marteinn)